Externalize analytics into custom hook
